+++ CHANGELOG.md
@@ -4,9 +4,17 @@ All notable changes to the "vscode-msr" extension will be documented in this fil
 
 Detail see `vscode-msr` extension [commit history](https://github.com/qualiu/vscode-msr/commits/master).
 
+## [1.0.2]
+
+- Reduced unimportant configuration keys.
+- Improved extra paths settings experience.
+- Support finding selected text other than just word.
+- Added more finding commands (Type `Ctrl + Shift + P` and then type `msr` to show these commands).
+- Use `cmd.exe` on Windows as default shell (`Powershell` requires extra escaping for characters like `'$'`).
+
 ## [1.0.1]
 
-- Updated configuration structure to support preciser searching.
+- Updated configuration structure to support more precise searching.
 - Refactored code to support:
   - Precise searching and flexibility.
   - Added finding commands (Type `Ctrl + Shift + P` and then type `msr` to show these commands).
@@ -19,4 +27,4 @@ Detail see `vscode-msr` extension [commit history](https://github.com/qualiu/vsc
 
 ## [0.0.1]
 
-- Initial release: Roughly support searching defintion and references for all type of coding languages.
+- Initial release: Roughly support searching definition and references for all type of coding languages.